+++ EKG_Project/Files/Data/Hardware/IECGSensor.cs
@@ -2,5 +2,5 @@ namespace EKG_Project;
 
 public interface IECGSensor
 {
-    int ReadRawSample();
+    double ReadRawSample();
 }
